+++ src/opts.c
@@ -53,7 +53,7 @@ static void usage(int err)
  * set_bool_opt:
  *      Set a boolean option to modify runtime behaviour.
  */
-bool set_bool_opt(unsigned opt, bool val)
+bool set_bool_opt(size_t opt, bool val)
 {
     switch (opt) {
     case VERBOSE_PRINT:
@@ -75,7 +75,7 @@ bool set_bool_opt(unsigned opt, bool val)
  * opt_is_set:
  *     Return boolean value for a runtime option.
  */
-bool opt_is_set(unsigned index)
+bool opt_is_set(size_t index)
 {
     return OPTIONS[index];
 }
